lai/requirements_windows_311.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
CairoSVG==2.7.1
diskcache==5.6.3
et_xmlfile==2.0.0
fastapi>=0.121.2
faster-whisper>=0.10
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx>=0.28.1
huggingface_hub==0.29.0
idna==3.11
Jinja2==3.1.6
llama_cpp_python==0.3.16
lxml==6.0.2
MarkupSafe==3.0.3
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
TTS==0.22.0
langdetect==1.0.9
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.18
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8
starlette>=0.49.3
transformers==4.41.2
tqdm==4.67.1
torch==2.5.1
torchaudio==2.5.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
fastmcp
mcp
